+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/pubsub/PubsubUnboundedSink.java
@@ -23,8 +23,11 @@ import com.google.protobuf.InvalidProtocolBufferException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ThreadLocalRandom;
 import org.apache.beam.sdk.coders.AtomicCoder;
@@ -69,7 +72,9 @@ import org.apache.beam.sdk.values.TypeDescriptor;
 import org.apache.beam.sdk.values.TypeDescriptors;
 import org.apache.beam.vendor.guava.v32_1_2_jre.com.google.common.annotations.VisibleForTesting;
 import org.apache.beam.vendor.guava.v32_1_2_jre.com.google.common.base.Preconditions;
+import org.apache.beam.vendor.guava.v32_1_2_jre.com.google.common.base.Strings;
 import org.apache.beam.vendor.guava.v32_1_2_jre.com.google.common.hash.Hashing;
+import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.joda.time.Duration;
 import org.joda.time.Instant;
@@ -201,8 +206,15 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
           break;
       }
 
+      // TODO(sjvanrossum): https://github.com/apache/beam/issues/31828
+      // NOTE: Null and empty ordering keys are treated as equivalent.
       @Nullable String topic = dynamicTopicFn.apply(element);
-      K key = keyFunction.apply(ThreadLocalRandom.current().nextInt(numShards), topic);
+      @Nullable String orderingKey = message.getOrderingKey();
+      int shard =
+          Strings.isNullOrEmpty(orderingKey)
+              ? ThreadLocalRandom.current().nextInt(numShards)
+              : Hashing.murmur3_32_fixed().hashString(orderingKey, StandardCharsets.UTF_8).asInt();
+      K key = keyFunction.apply(shard, topic);
       o.output(KV.of(key, OutgoingMessage.of(message, timestampMsSinceEpoch, recordId, topic)));
     }
 
@@ -219,6 +231,16 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
 
   /** Publish messages to Pubsub in batches. */
   private static class WriterFn extends DoFn<Iterable<OutgoingMessage>, Void> {
+    private class OutgoingData {
+      int bytes;
+      List<OutgoingMessage> messages;
+
+      OutgoingData() {
+        this.bytes = 0;
+        this.messages = new ArrayList<>(publishBatchSize);
+      }
+    }
+
     private final PubsubClientFactory pubsubFactory;
     private final @Nullable ValueProvider<TopicPath> topic;
     private final String timestampAttribute;
@@ -305,27 +327,45 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
     }
 
     @ProcessElement
+    @SuppressWarnings("ReferenceEquality")
     public void processElement(ProcessContext c) throws Exception {
-      List<OutgoingMessage> pubsubMessages = new ArrayList<>(publishBatchSize);
-      int bytes = 0;
+      // TODO(sjvanrossum): Refactor the write transform so this map can be indexed with topic +
+      // ordering key and have bundle scoped lifetime.
+      // NOTE: A single publish request may only write to one ordering key.
+      // See https://cloud.google.com/pubsub/docs/publisher#using-ordering-keys for details.
+      Map<String, OutgoingData> orderingKeyBatches = new HashMap<>();
+      @MonotonicNonNull String currentOrderingKey = null;
+      @Nullable OutgoingData currentBatch = null;
       for (OutgoingMessage message : c.element()) {
-        if (!pubsubMessages.isEmpty()
-            && bytes + message.getMessage().getData().size() > publishBatchBytes) {
+        String messageOrderingKey = message.getMessage().getOrderingKey();
+        if (currentOrderingKey == null || !currentOrderingKey.equals(messageOrderingKey)) {
+          currentOrderingKey = messageOrderingKey;
+          currentBatch = orderingKeyBatches.get(currentOrderingKey);
+        }
+        if (currentBatch == null) {
+          currentBatch = new OutgoingData();
+          orderingKeyBatches.put(currentOrderingKey, currentBatch);
+        } else if (currentBatch.bytes + message.getMessage().getData().size() > publishBatchBytes) {
+          // TODO(sjvanrossum): https://github.com/apache/beam/issues/31800
+
           // Break large (in bytes) batches into smaller.
           // (We've already broken by batch size using the trigger below, though that may
           // run slightly over the actual PUBLISH_BATCH_SIZE. We'll consider that ok since
           // the hard limit from Pubsub is by bytes rather than number of messages.)
           // BLOCKS until published.
-          publishBatch(pubsubMessages, bytes);
-          pubsubMessages.clear();
-          bytes = 0;
+          publishBatch(currentBatch.messages, currentBatch.bytes);
+          currentBatch.messages.clear();
+          currentBatch.bytes = 0;
         }
-        pubsubMessages.add(message);
-        bytes += message.getMessage().getData().size();
+        currentBatch.messages.add(message);
+        // TODO(sjvanrossum): https://github.com/apache/beam/issues/31800
+        currentBatch.bytes += message.getMessage().getData().size();
       }
-      if (!pubsubMessages.isEmpty()) {
-        // BLOCKS until published.
-        publishBatch(pubsubMessages, bytes);
+      for (OutgoingData batch : orderingKeyBatches.values()) {
+        if (!batch.messages.isEmpty()) {
+          // BLOCKS until published.
+          publishBatch(batch.messages, batch.bytes);
+        }
       }
     }
 
@@ -378,6 +418,12 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
    */
   private final int numShards;
 
+  /**
+   * Publish messages with an ordering key. Currently unsupported with DataflowRunner's Pubsub sink
+   * override.
+   */
+  private final boolean publishBatchWithOrderingKey;
+
   /** Maximum number of messages per publish. */
   private final int publishBatchSize;
 
@@ -402,6 +448,7 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
       String timestampAttribute,
       String idAttribute,
       int numShards,
+      boolean publishBatchWithOrderingKey,
       int publishBatchSize,
       int publishBatchBytes,
       Duration maxLatency,
@@ -412,6 +459,7 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
     this.timestampAttribute = timestampAttribute;
     this.idAttribute = idAttribute;
     this.numShards = numShards;
+    this.publishBatchWithOrderingKey = publishBatchWithOrderingKey;
     this.publishBatchSize = publishBatchSize;
     this.publishBatchBytes = publishBatchBytes;
     this.maxLatency = maxLatency;
@@ -424,13 +472,15 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
       ValueProvider<TopicPath> topic,
       String timestampAttribute,
       String idAttribute,
-      int numShards) {
+      int numShards,
+      boolean publishBatchWithOrderingKey) {
     this(
         pubsubFactory,
         topic,
         timestampAttribute,
         idAttribute,
         numShards,
+        publishBatchWithOrderingKey,
         DEFAULT_PUBLISH_BATCH_SIZE,
         DEFAULT_PUBLISH_BATCH_BYTES,
         DEFAULT_MAX_LATENCY,
@@ -444,6 +494,7 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
       String timestampAttribute,
       String idAttribute,
       int numShards,
+      boolean publishBatchWithOrderingKey,
       String pubsubRootUrl) {
     this(
         pubsubFactory,
@@ -451,6 +502,7 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
         timestampAttribute,
         idAttribute,
         numShards,
+        publishBatchWithOrderingKey,
         DEFAULT_PUBLISH_BATCH_SIZE,
         DEFAULT_PUBLISH_BATCH_BYTES,
         DEFAULT_MAX_LATENCY,
@@ -464,6 +516,7 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
       String timestampAttribute,
       String idAttribute,
       int numShards,
+      boolean publishBatchWithOrderingKey,
       int publishBatchSize,
       int publishBatchBytes) {
     this(
@@ -472,6 +525,7 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
         timestampAttribute,
         idAttribute,
         numShards,
+        publishBatchWithOrderingKey,
         publishBatchSize,
         publishBatchBytes,
         DEFAULT_MAX_LATENCY,
@@ -485,6 +539,7 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
       String timestampAttribute,
       String idAttribute,
       int numShards,
+      boolean publishBatchWithOrderingKey,
       int publishBatchSize,
       int publishBatchBytes,
       String pubsubRootUrl) {
@@ -494,6 +549,7 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
         timestampAttribute,
         idAttribute,
         numShards,
+        publishBatchWithOrderingKey,
         publishBatchSize,
         publishBatchBytes,
         DEFAULT_MAX_LATENCY,
@@ -518,6 +574,10 @@ public class PubsubUnboundedSink extends PTransform<PCollection<PubsubMessage>, 
   /** Get the id attribute. */
   public @Nullable String getIdAttribute() {
     return idAttribute;
+  }
+
+  public boolean getPublishBatchWithOrderingKey() {
+    return publishBatchWithOrderingKey;
   }
 
   @Override